# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Nihongo Learning API",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/learning-sessions/{session_id}/questions": {
      "get": {
        "tags": [
          "Learning Sessions"
        ],
        "summary": "Get Session Questions",
        "operationId": "Learning Sessions-get_session_questions_api_v1_learning-sessions_session_id_questions",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Session Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/SessionQuestionResponse"
                  },
                  "title": "Response Learning Sessions-Get Session Questions Api V1 Learning-Sessions Session Id Questions"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/admin/questions/": {
      "get": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Get Questions",
        "operationId": "Admin Questions-get_questions_api_v1_admin_questions_",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/QuestionResponseAdmin"
                  },
                  "type": "array",
                  "title": "Response Admin Questions-Get Questions Api V1 Admin Questions "
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Create Question",
        "operationId": "Admin Questions-create_question_api_v1_admin_questions_",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ]
      }
    },
    "/api/v1/admin/questions/{question_id}": {
      "get": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Get Question",
        "operationId": "Admin Questions-get_question_api_v1_admin_questions_question_id",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "patch": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Update Question",
        "operationId": "Admin Questions-update_question_api_v1_admin_questions_question_id",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{question_id}/submit-review": {
      "post": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Submit Review",
        "operationId": "Admin Questions-submit_review_api_v1_admin_questions_question_id_submit-review",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{question_id}/approve": {
      "post": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Approve Question",
        "operationId": "Admin Questions-approve_question_api_v1_admin_questions_question_id_approve",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionReviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{question_id}/reject": {
      "post": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Reject Question",
        "operationId": "Admin Questions-reject_question_api_v1_admin_questions_question_id_reject",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionReviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{question_id}/request-revision": {
      "post": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Request Revision",
        "operationId": "Admin Questions-request_revision_api_v1_admin_questions_question_id_request-revision",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QuestionReviewCreate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{question_id}/publish": {
      "post": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Publish Question",
        "operationId": "Admin Questions-publish_question_api_v1_admin_questions_question_id_publish",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{question_id}/unpublish": {
      "post": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Unpublish Question",
        "operationId": "Admin Questions-unpublish_question_api_v1_admin_questions_question_id_unpublish",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{question_id}/archive": {
      "post": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Archive Question",
        "operationId": "Admin Questions-archive_question_api_v1_admin_questions_question_id_archive",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QuestionResponseAdmin"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/admin/questions/{question_id}/history": {
      "get": {
        "tags": [
          "Admin Questions"
        ],
        "summary": "Get Question History",
        "operationId": "Admin Questions-get_question_history_api_v1_admin_questions_question_id_history",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "question_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Question Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/QuestionRevisionResponse"
                  },
                  "title": "Response Admin Questions-Get Question History Api V1 Admin Questions Question Id History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "QuestionCreate": {
        "properties": {
          "lesson_id": {
            "type": "string",
            "title": "Lesson Id"
          },
          "reading_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Id"
          },
          "audio_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Asset Id"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty",
            "default": 1
          },
          "prompt_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Prompt Json"
          },
          "explanation_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation Json"
          },
          "answer_key_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Key Json"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "question_type",
          "skill",
          "prompt_json"
        ],
        "title": "QuestionCreate"
      },
      "QuestionResponseAdmin": {
        "properties": {
          "lesson_id": {
            "type": "string",
            "title": "Lesson Id"
          },
          "reading_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reading Id"
          },
          "audio_asset_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Asset Id"
          },
          "question_type": {
            "$ref": "#/components/schemas/QuestionType"
          },
          "skill": {
            "$ref": "#/components/schemas/SkillType"
          },
          "difficulty": {
            "type": "integer",
            "title": "Difficulty",
            "default": 1
          },
          "prompt_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Prompt Json"
          },
          "explanation_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation Json"
          },
          "id": {
            "type": "string",
            "title": "Id"
          },
          "answer_key_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Key Json"
          },
          "status": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "is_ai_generated": {
            "type": "boolean",
            "title": "Is Ai Generated"
          },
          "version_number": {
            "type": "integer",
            "title": "Version Number"
          },
          "created_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created By"
          },
          "reviewed_by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewed By"
          }
        },
        "type": "object",
        "required": [
          "lesson_id",
          "question_type",
          "skill",
          "prompt_json",
          "id",
          "status",
          "is_ai_generated",
          "version_number",
          "created_by",
          "reviewed_by"
        ],
        "title": "QuestionResponseAdmin"
      },
...
      "QuestionReviewCreate": {
        "properties": {
          "status_given": {
            "$ref": "#/components/schemas/QuestionStatus"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "status_given"
        ],
        "title": "QuestionReviewCreate"
      },
...
      "QuestionType": {
        "type": "string",
        "enum": [
          "MULTIPLE_CHOICE",
          "TRUE_FALSE",
          "MATCHING",
          "ORDERING",
          "CLOZE",
          "LISTENING",
          "READING"
        ],
        "title": "QuestionType"
      },
      "QuestionUpdate": {
        "properties": {
          "prompt_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Json"
          },
          "answer_key_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Answer Key Json"
          },
          "explanation_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Explanation Json"
          },
          "difficulty": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Difficulty"
          }
        },
        "type": "object",
        "title": "QuestionUpdate"
      },
...
      "SessionQuestionResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "session_id": {
            "type": "string",
            "title": "Session Id"
          },
          "order_number": {
            "type": "integer",
            "title": "Order Number"
          },
          "is_answered": {
            "type": "boolean",
            "title": "Is Answered"
          },
          "is_correct": {
            "type": "boolean",
            "title": "Is Correct"
          },
          "score": {
            "type": "integer",
            "title": "Score"
          },
          "user_answer_json": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "User Answer Json"
          },
          "question": {
            "$ref": "#/components/schemas/QuestionResponseLearner"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session_id",
          "order_number",
          "is_answered",
          "is_correct",
          "score",
          "question"
        ],
        "title": "SessionQuestionResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/login"
          }
        }
      }
    }
  }
}